lib/customer_flight_info/consumer/requirements.txt
pandas==2.2.2
requests==2.31.0
tomli==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pika==1.3.2
